requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.2
fsspec==2024.3.1
idna==3.6
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
regex==2023.12.25
requests==2.31.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
